Type Definition rowdy_diesel::postgres::Authenticator

type Authenticator = Authenticator<PgConnection>;

A rowdy authenticator that uses a PostgresSQL backed database to provide the users

Methods

impl Authenticator

pub fn with_uri(uri: &str) -> Result<Self, Error>

Using a database connection string of the form postgresql://[user[:password]@][host][:port][/database_name], create an authenticator that is backed by a connection pool to a PostgresSQL database

pub fn with_configuration(
host: &str,
port: u16,
database: &str,
user: &str,
pass: &str
) -> Result<Self, Error>

Create a new Authenticator with a database config