Module rowdy_diesel::schema

Schema of the users table that will be used with Rowdy If you have more sophisticated needs, you are able to add more columns to the basic columns needed for rowdy to work.

The schema required for the table as expressed using Diesel's table! macro is

table! {
    users (username) {
        username -> Varchar,
        hash -> Binary,
        salt -> Varbinary,
    }
}

In standard SQL parlance, this is equivalent to

CREATE TABLE IF NOT EXISTS `users` (
    `username` VARCHAR(255) UNIQUE NOT NULL,
    `hash` BINARY(32) NOT NULL,
    `salt` VARBINARY(255) NOT NULL,
    PRIMARY KEY (`username`)
);

Modules

users
Table used to hold users and their hashed passwords

Traits

Migration
Trait to provide idempotent minimal migration to create the table necessary for rowdy-diesel to work. If you have more sophisticated needs, you are able to add more columns to the basic columns needed for rowdy to work.