Trait rowdy_diesel::schema::Migration

pub trait Migration<T> where
    T: Connection + 'static, {
    type Connection: Deref<Target = T>;
    fn connection(&self) -> Result<Self::Connection, Error>;

    fn migration_query(&self) -> &str;

    fn migrate(&self) -> Result<(), Error> { ... }
}

Trait to provide idempotent minimal migration to create the table necessary for rowdy-diesel to work. If you have more sophisticated needs, you are able to add more columns to the basic columns needed for rowdy to work.

Associated Types

type Connection: Deref<Target = T>

Connection type for the migration to work with

Required Methods

fn connection(&self) -> Result<Self::Connection, Error>

Provide a connection for the migration to work with

fn migration_query(&self) -> &str

Format the migration query based on the connection type

Provided Methods

fn migrate(&self) -> Result<(), Error>

Provide idempotent minimal migration to create the table necessary for rowdy-diesel to work