Module rowdy_diesel::schema::users

Table used to hold users and their hashed passwords

Hashing is done with the argon2i algorithm with a salt.

Modules

columns
Contains all of the columns of this table

dsl
Re-exports all of the columns of this table, as well as the table struct renamed to the module name. This is meant to be glob imported for functions which only deal with one table.

Structs

hash
A argon2i hash of the user's password

salt
Salt used to generate the password hash

star
Represents table_name.*, which is sometimes needed for efficient count queries. It cannot be used in place of all_columns, and has a SqlType of () to prevent it being used that way

table
The actual table struct

username
Username for the user. Also the primary key

Constants

all_columns
A tuple of all of the columns on this table

Type Definitions

BoxedQuery
Helper type for representing a boxed query from this table

SqlType
The SQL type of all of the columns on this table