Struct rowdy_diesel::Authenticator

pub struct Authenticator<T> where
    T: Connection + 'static,  { /* fields omitted */ }

A generic authenticator backed by a connection to a database via diesel.

Instead of using this, you should use the "specialised" authenticators defined in the mysql, pg, or sqlite modules for your database.

Passwords are hasahed with argon2i, in addition to a salt.

Methods

impl<T> Authenticator<T> where
T: Connection + 'static,
String: FromSql<Text, <T as Connection>::Backend>,
Vec<u8>: FromSql<Binary, <T as Connection>::Backend>,

pub fn hash_password(password: &str, salt: &[u8]) -> Result<String, Error>

Hash a password with the salt. See struct level documentation for the algorithm used.

pub fn verify(
&self,
username: &str,
password: &str,
include_refresh_payload: bool
) -> Result<AuthenticationResult, Error>

Verify that some user with the provided password exists in the database, and the password is correct.

Returns the payload to be included in a refresh token if successful